pyproject.toml
[tool.poetry]
name = "kroger-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.7"
uvicorn = "^0.34.0"
requests = "^2.32.3"
pymongo = "^4.11"
python-dotenv = "^1.0.1"
sentence-transformers = "^3.4.1"
faiss-cpu = "^1.10.0"
motor = "^3.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
